% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.2.bed.R
\name{bam.2.bed}
\alias{bam.2.bed}
\title{Convert bam files to bed files.}
\usage{
bam.2.bed(bamFiles, paired = FALSE, postsort = TRUE, cigar = FALSE,
  threads = getOption("threads", 1L), sortBuffer = "1G",
  sortThreads = NULL)
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{paired}{A boolean indicating if bed intervals should be created with the inferred fragment derived from mate-pair alignments.}

\item{postsort}{A boolean indicating if the bed file should be sorted upon creation.}

\item{cigar}{A boolean indicating whether to use CIGAR strings as interval names.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bam.2.bed} is a wrapper for bedtools bamtobed.
}

