% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.2.sam.R
\name{bam.2.sam}
\alias{bam.2.sam}
\title{Convert bam files to sam files}
\usage{
bam.2.sam(bamFiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bam.2.sam} is a simple wrapper for samtools view to create sam files from bam files.
}

