% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.cat.R
\name{bam.cat}
\alias{bam.cat}
\title{Concatenate bam files}
\usage{
bam.cat(bamFiles, outname)
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{outname}{A vector of length one specifying the concatenated file name.}
}
\description{
\code{bam.cat} is a simple wrapper for samtools cat for combining bam files.
}

