% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.count.R
\name{bam.count}
\alias{bam.count}
\title{Count reads in bam files}
\usage{
bam.count(bamfiles, q = 0, threads = getOption("threads", 1L))
}
\arguments{
\item{q}{An integer >= 0 specifying the minimum quality of a read for it to be counted.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}

\item{bamFiles}{A character vector of paths to bam files.}
}
\description{
\code{bam.count} is a simple wrapper for \code{samtools view -c} for counting the number of reads in a bam file.
}

