% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.index.R
\name{bam.index}
\alias{bam.index}
\title{Index bam files}
\usage{
bam.index(bamFiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bam.index} is a simple wrapper for \code{samtools index} for indexing bam files.
}

