% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bam.merge.R
\name{bam.merge}
\alias{bam.merge}
\title{Merge sorted bam files}
\usage{
bam.merge(bamFiles, outname, threads = getOption("threads", 1L))
}
\arguments{
\item{bamFiles}{A character vector of paths to bam files.}

\item{outname}{A vector of length one specifying the concatenated file name.}

\item{threads}{An integer specifying the number of BAM compression threads to use}
}
\description{
\code{bam.merge} is a simple wrapper for samtools merge for merging sorted bam files.
}

