% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bb.2.bed.R
\name{bb.2.bed}
\alias{bb.2.bed}
\title{Convert bigBed files to bed files}
\usage{
bb.2.bed(datafiles, threads = getOption("threads", 1L))
}
\arguments{
\item{datafiles}{A character vector of paths to bigBed files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bb2bed} bb2bed converts bigBed files to bed files with \code{bigBedToBed}.
}

