% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.2.bb.R
\name{bed.2.bb}
\alias{bed.2.bb}
\title{Convert bed files to bigBed files}
\usage{
bed.2.bb(datafiles, genomefile, threads = getOption("threads", 1L))
}
\arguments{
\item{datafiles}{A character vector of paths to bed files.}

\item{genomefile}{A string specifying the path of the file with chromosome sizes}

\item{threads}{A positive integer specifying how many files to process simultaneously.}
}
\description{
\code{bed2bbb} converts bed files to bigBed files with \code{bedToBigBed}.
}

