% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.2.bg.R
\name{bed.2.bg}
\alias{bed.2.bg}
\title{Convert bed files to bedGraph files}
\usage{
bed.2.bg(bedfiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bedfiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many files to process simultaneously.}
}
\description{
\code{bed2bg} converts bed files to bedGraph files by printing bed columns 1,2,3,5.
}

