% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.hist.R
\name{bed.hist}
\alias{bed.hist}
\title{Plot a histogram of interval sizes in bed files}
\usage{
bed.hist(bedFiles, threads = getOption("threads", 1L), ...)
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many beds to process simultaneously.}
}
\description{
\code{bed.hist} plots histograms, density functions, or CDFs of interval sizes in bed files.
}

