% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.parselengths.R
\name{bed.parselengths}
\alias{bed.parselengths}
\title{Parse intervals in bed files by size.}
\usage{
bed.parselengths(bedFiles, breaks, threads = getOption("threads", 1L))
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{breaks}{A numeric vector of read lengths used as break points to parse reads by length. For example, if breaks=c(0,100,200), reads will be parsed into reads of length 0-99 and 100-199. Length of vector must beat least 2.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bed.parselengths} separates intervals from a bed file into separate bed files based on interval sizes using the system \code{awk}.
}

