% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.sizes.R
\name{bed.sizes}
\alias{bed.sizes}
\title{Retreive interval sizes in bed files}
\usage{
bed.sizes(bedFiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bed.sizes} returns a list of vectors containing the sizes for intervals in a set of bed files.
}

