% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bed.sort.R
\name{bed.sort}
\alias{bed.sort}
\title{Sort beds by chromosome and position}
\usage{
bed.sort(bedFiles, threads = getOption("threads", 1L), sortBuffer = "1G",
  sortThreads = NULL)
}
\arguments{
\item{bedFiles}{A character vector of paths to bed files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}

\item{sortThreads}{A positive integer specifying the number of sorting and compression threads}

\item{memory}{String specifying maximum memory per thread; suffix K/M/G recognized.}

\item{sortByName}{Boolean. If TRUE, reads are sorted by name. If FALSE, reads are sorted by chromosome/position}
}
\description{
\code{bed.sort} sorts bed files chromosome then position using the system's \code{sort}.
}

