% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg.2.bw.R
\name{bg.2.bw}
\alias{bg.2.bw}
\title{Convert bedGraph files to bigWig files}
\usage{
bg.2.bw(datafiles, genomefile, threads = getOption("threads", 1L))
}
\arguments{
\item{datafiles}{A character vector of paths to bedGraph files.}

\item{genomefile}{A string specifying the path of the file with chromosome sizes}

\item{threads}{A positive integer specifying how many files to process simultaneously.}
}
\description{
\code{bg.2.bw} converts bedGraph files to bigWig files with \code{bedGraphToBigWig}.
}

