% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg.hist.R
\name{bg.hist}
\alias{bg.hist}
\title{Plot a histogram of scores sizes in bedGraph files}
\usage{
bg.hist(bgfiles, threads = getOption("threads", 1L), ...)
}
\arguments{
\item{threads}{A positive integer specifying how many bedGraph files to process simultaneously.}

\item{bgFiles}{A character vector of paths to bedgraph files.}
}
\description{
\code{bg.hist} plots histograms, density functions, or CDFs of scores in bedGraph files.
}

