% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/bg.scores.R
\name{bg.scores}
\alias{bg.scores}
\title{Retreive scores in bedGraph files}
\usage{
bg.scores(bgFiles, threads = getOption("threads", 1L))
}
\arguments{
\item{bgFiles}{A character vector of paths to bedgraph files.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{bg.scores} returns a list of vectors containing the scores for intervals in a set of bedGraph files.
}

