% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/mat.make.R
\name{mat.make}
\alias{mat.make}
\title{Create a matrix of data aligned at a set of features.}
\usage{
mat.make(scorefiles, features, featurenames = NULL, regionsize = 2000,
  windowsize = 10, strand = FALSE, featurecenter = TRUE, start = 2,
  stop = 3, prunefeaturesto = NULL, narrowpeak = FALSE, maskbed = NULL,
  bgfiller = 0, prunescores = FALSE, rpm = TRUE, closest = NULL,
  cores = "max", meta = FALSE, metaflank = 1000, suffix = NULL,
  scoremat = TRUE, fragmats = FALSE)
}
\arguments{
\item{scorefiles}{A character vector of paths to data files that are used to fill the matrix. Can be bed, bedGraph, wiggle, bigWig. Can also include URLs and/or gzipped forms of these files.}

\item{features}{A character vector of paths to genomic intervals corresponding to rows in the matrix that are used to align values in 'scorefiles'.}

\item{featurenames}{A character vector equal in length to 'features' that is used as a suffix in the output file names. If NULL, uses the file names of 'features'.}

\item{regionsize}{A positive integer specifying what size (bp) interval around each feature is used to generate the matrix. Must be divisible by 'windowsize'.}

\item{windowsize}{A positive integer specifying what size (bp) windows are used to generate the matrix.}

\item{strand}{Boolean indicating if the orientation of the scores around minus-stranded features should be reversed in the matrix.}

\item{featurecenter}{Boolean indicating if the matrix should be aligned to the midpoint of the intervals in 'features'. If FALSE, features will be aligned to the coordinate specified by 'start' (or 'stop' for minus-stranded features if 'strand' is TRUE).}

\item{start}{A positive integer specifying the column to be used in 'features' as the 5' end of features. Can be set to '7' for aligning at the start codons of genes in bed12 format, '3' for aligning at the end of genes, or '8' for aligning at stop codons.}

\item{stop}{A positive integer specifying the column to be used in 'features' as the 3' end of features.}

\item{prunefeaturesto}{A string specifying a bed file to filter intervals in 'features' by. Only invervals in 'features' that overlap with intervals in 'prunefeaturesto' will be used to construct the matrix. Useful if data in 'scorefiles' only exists for a proportion of the genome and wish to limit features used to construct the matrix to those regions.}

\item{narrowpeak}{Boolean indicating if the matrix are aligned to summits of peaks in a narrowpeak file that is present in 'features'. Only used if summits are specified in the narrowpeak file and if 'featurecenter' is TRUE.}

\item{maskbed}{A string specifying a bed file to filter values in the matrix by. If TRUE, only windows in the matrix that overlap with intervals in 'maskbed' will be assigned scores, and all other windows will contain 'NA'. Useful for preventing the FALSE assignement of zeroes in regions that have no value in 'scorefiles', such as in data from sequence-capture experiments.}

\item{bgfiller}{A value to be used for windows in the matrix that have no score in bedGraph files specified in 'scorefiles'. Can be a number or NA.}

\item{prunescores}{Boolean indicating if scores are intersected with the locations of windows in the matrix prior to generating the matrix. Does not affect or change the resulting matrix but does speed up the function for very large files in 'scoresfiles'.}

\item{rpm}{Boolean indicating if interval densities calculated from bed files specified in 'scorefiles' are normalized to reads-per-million (RPM). If TRUE, bed interval counts in each window is multiplied by 1000000/(# of intervals in bed file). Useful if using reads with varying amounts in 'scorefiles' and a comparison of resulting matrices is desired.}

\item{closest}{A string specifying a bed file to label features in the matrix. This will not affect the values in matrices but will change the names for each feature in the matrix based on the closest interval in 'closest'. Useful for assigning gene names to ChIP peaks used as 'features'.}

\item{cores}{A positive integer specifying how many bams to process simultaneously.}

\item{meta}{Boolean indicating if 'metafeatures' are used to generate the matrix. If FALSE, matrices are aligned to single coordinates, and all windows in the matrix represent a fixed-sized window surrounding the coordinates. If TRUE, matrices are aligned to both the 'start' and 'stop' coordinates, and the distances and windows between the start and stop are scaled to a fixed value specified in 'regionsize'. Useful for generating a matrix of values within and around varying-size genes.}

\item{metaflank}{A positive integeger specifying the distance from metafeature boundaries used to generate the matrix. Must be divisible by windowsize.}

\item{suffix}{A string that is appended to the output matrix names.}

\item{scoremat}{Boolean indicating if score matrices are generated relative to features.}

\item{fragmats}{Boolean indicating if a special type of matrix, a fragment-size matrix, is generated using the sizes of intervals for bed files in 'scorefiles'. Such matrices can be used to generate 'v-plots'.}
}
\description{
\code{mat.make} creates a matrix of scores aligned at a set of genomic intervals, where each row is an interval and each column is a window surrounding each interval.
}

