% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/sam.2.bam.R
\name{sam.2.bam}
\alias{sam.2.bam}
\title{Convert sam files to bam files}
\usage{
sam.2.bam(samFiles, minQual = NULL, threads = getOption("threads", 1L))
}
\arguments{
\item{samFiles}{A character vector of paths to sam files.}

\item{minQual}{A positive integer specifying the minimum alignment quality a read needs to be written.}

\item{threads}{A positive integer specifying how many bams to process simultaneously.}
}
\description{
\code{sam.2.bam} is a simple wrapper for samtools view to create bam files from sam files.
}

